Move tests using sys/mount to not build on Windows
